refactor: 유지보수 향상을 위한 모달 컴포넌트 분리
