+++ components/src/types/modal.ts
@@ -1,0 +1,2 @@
+export type ModalPosition = "center" | "bottom";
+export type ModalSize = "small" | "medium" | "large";
